fix: change totalPrice to DECIMAL

+++ back-end/src/database/migrations/20220427193919-create-sales.js
@@ -18,7 +18,7 @@ module.exports = {
       },
       total_price: {
         allowNull: false,
-        type: Sequelize.FLOAT
+        type: Sequelize.DECIMAL(10, 2),
       },
       delivery_address: {
         allowNull: false,
